PHP bugfixes and add func for admin
	* buggfix correct redirection for all comment functions.
	* add new setcookie('userType') to login function.
	* add || to edit and delete statemens regardning
	  comments/blogposts so that admin users also can edit and
	  delete.
	* bugfix all comments asosiated with a blogpost are now also
	  deleted when the blogpost is deleted.
	* some smaler changes to the way errormesseges are displayed.

+++ src/Models/BlogpostModel.php
@@ -261,6 +261,16 @@ class BlogpostModel extends AbstractModel
         
         try {
 
+            $query = 'DELETE FROM blogposts_comments WHERE post_id =:id';
+            $statement = $this->db->prepare($query);
+            $statement->bindValue("id", $id);
+            
+            if(!$statement->execute()) 
+            {
+                throw new Exception($statement->errorinfo()[2]);
+            }
+    
+
             $query = 'DELETE FROM blogposts_content WHERE id =:id';
             $statement = $this->db->prepare($query);
             $statement->bindValue("id", $id);
